# Improve multipart parser performance (from quadratic to linear)