--- a/Tests/MultipartTests/ParserTests.swift
+++ b/Tests/MultipartTests/ParserTests.swift
@@ -12,6 +12,8 @@ class ParserTests: XCTestCase {
         ("testHeaders", testHeaders),
         ("testEpilogue", testEpilogue),
         ("testFormData", testFormData),
+        ("testExtractBoundary", testExtractBoundary),
+        ("testExtractBoundaryWithQuotes", testExtractBoundaryWithQuotes),
     ]
 
     func testInit() throws {