@ole 2018-03-13
So the tests on Linux are failing. I'm not sure if this is because I made a mistake or if there's a problem with XCTest's performance measurement APIs on Linux. Either way, I'd like to get some input on your current performance testing setup before I invest time in fixing this. After all, there's no point in fixing it if you prefer something other than XCTest-based performance tests.