@ole 2018-03-13
Independent of the performance tests, I'll have to fix the Swift 3 error because I used a Swift-4-only feature. I'll send a commit for that.