@ole 2018-03-14
@tanner0101 Great. The tests are passing now (and I added two previously missing tests to the Linux test suite in the process). Let me know if you need anything else from me.